fix(Graph): 🪲️nodes without in/out pins caused crash
Now there aren't nodes with no in/out, but even if there were, they
would be drawn anyway.

+++ src/include/engine/node_graph.h
@@ -9,7 +9,7 @@
 #include "engine/stream.h"
 
 struct Link {
-  uint64_t id{};
+  int id{};
   uint64_t start_pin_id{};
   uint64_t end_pin_id{};
 };
